Patch users `paths` file in compilerOptions (#2620)
This is something I spent very long trying to avoid, but in area-code
there is no configuration that either @groy-514 or I could use that
doesnt rely on using `paths` in tsconfig. When we copy over and pnpm
deploy, we lose these references. This only handles one specific case of
a hard problem and what to do with more external dependencies