chore: cleaning up internal app (#2622)
This pull request removes the entire GitHub Actions workflow for
deploying the internal app. This means that automatic deployments and
related notifications for the internal app via this workflow will no
longer occur.

Workflow removal:

* Deleted `.github/workflows/deploy-internal-app.yaml`, which contained
all steps for building, pushing, and deploying the internal app, as well
as Slack notifications on failure.